requirements.txt
# Installing collected packages: urllib3, typing-extensions, tenacity, sniffio, PyYAML, packaging, orjson, msgpack, jsonpointer, idna, h11, exceptiongroup, charset-normalizer, certifi, annotated-types, requests, pydantic-core, jsonpatch, httpcore, anyio, requests-toolbelt, pydantic, httpx, langsmith, langgraph-sdk, langchain-core, langgraph-checkpoint, langgraph
# Successfully installed PyYAML-6.0.2 annotated-types-0.7.0 anyio-4.6.2.post1 certifi-2024.8.30 charset-normalizer-3.4.0 exceptiongroup-1.2.2 h11-0.14.0 httpcore-1.0.7 httpx-0.28.0 idna-3.10 jsonpatch-1.33 jsonpointer-3.0.0 langchain-core-0.3.21 langgraph-0.2.56 langgraph-checkpoint-2.0.8 langgraph-sdk-0.1.43 langsmith-0.1.147 msgpack-1.1.0 orjson-3.10.12 packaging-24.2 pydantic-2.10.3 pydantic-core-2.27.1 requests-2.32.3 requests-toolbelt-1.0.0 sniffio-1.3.1 tenacity-9.0.0 typing-extensions-4.12.2 urllib3-2.2.3  Installing collected packages: propcache, numpy, multidict, greenlet, frozenlist, attrs, async-timeout, aiohappyeyeballs, yarl, SQLAlchemy, aiosignal, aiohttp, langchain-text-splitters, langchain
#Successfully installed SQLAlchemy-2.0.36 aiohappyeyeballs-2.4.4 aiohttp-3.11.9 aiosignal-1.3.1 async-timeout-4.0.3 attrs-24.2.0 frozenlist-1.5.0 greenlet-3.1.1 langchain-0.3.9 langchain-text-splitters-0.3.2 multidict-6.1.0 numpy-1.26.4 propcache-0.2.1 yarl-1.18.3

python-dotenv
langchain
langgraph
langchain-openai
langchain-anthropic
langchain-community
langchain-core

